Make the content integrity checker more extendable
The integrity checker now depends on an abstract class that contains the
base code that is necessary to perform the check on a storage.

This abstract class can be inherited by concret classes that will define
the checker behavior when a corrupted or missing content is find
(Note that for now, the checker runs on a local storage so the list of
content to check is made from walking the filesystem. So a missing
content shouldn't happend. But another version could get this list from
another source that is not the filesystem).