Add an ObjStorage.list_content() method
This method generates all the obj_ids in a storage, with support for
pagination so it does not bloat too much a beefy objstorage.
The support for pagination implies making the iteration over an
ObjStorage sorted.

This method is exposed via the RPC using the /content endpoint.

Add some tests.

Note: we skip the test_list_content for the StripingObjStorage since we
have no way to implement pagination support for the list_content()
method in this backend.