azure: Specialize get_batch() to fetch objects concurrently
The default implementation is a naive 'for' loop calling 'get'
sequentially.

This replaces this default implementation with a local asyncio loop,
which manages a task for each of the blobs to fetch.

To avoid code duplication, this also rewrites 'get' to use the same
code as 'get_batch'. Quick benchmarks did not show a noticeable
slowdown when doing this.