client: Do not depend on how the server chunks the response
Currently, the server uses HTTP chunked encoding to send object ids
one by one, each in its own HTTP frame.

I think this is a mistake to rely on such a detail of the HTTP protocol
in a high-level API like this.

Additionally, a future commit will rewrite the server to use Flask
instead of aiohttp, which does not allow this kind of fine-grained
control about HTTP chunks.