api/server: Fix streaming responses implementation
Do not yield an empty byte in list_content as it triggers an error
since the release of werkzeug 2.1.0 which bumps HTTP protocol version
from 1.0 to 1.1.

Ensure objstorage is initialized in streaming responses implementation,
if no POST requests on other endpoints have been issued the global
variable objstorage was still equal to None.

Closes T4119