azure: Add tests based on Azurite in addition to mocks
In addition to add support for multi-hashes, we will need to use more advanced
features of the Azure Blob Storage, so using only mock-based tests will be
less reliable.

This commit adds the option to run the test suite with Azurite, a clone
of the Azure Blob Storage API, which is a standard tool used in tests.