winery: properly reset SharedBase.locked_shard when appropriate
When a shard is moved to an unlocked state, the state of the SharedBase
object should be updated to match.

When a shard is set to STANDBY or FULL (from WRITING), release it from
the SharedBase. This forces shard_packing_starts to receive the shard
name, as a SharedBase can't have a locked_shard in state FULL set
anymore.