winery: introduce a new CLEANING shard state
The CLEANING state is between PACKED and READONLY, and records the
cleanup of the Read-Write shard.

To allow implementing the CLEANING state, add a list of hosts that are
known to have mapped the image read-only. This list is only advisory,
and enables moving from the PACKED to the READONLY state and dropping
the read-write shard once all readers have noticed that the shard has
been packed.