Implement delete() in WineryObjStorage
We now allow to delete objects from Winery. The object is first marked
as deleted in the shared base. This makes it inaccessible from further
`get()` calls. If the object still in a RWShard, the relevant line in
the `objects` table is deleted as well.

To take care of removing the data in ROShards, a dedicated host able to
map images in read-write mode can then call:

    swh objstorage winery clean-deleted-objects

This will zero out any deleted object present in the images and remove
their key from the indices. Object keys will then be removed from the
shared database.

Thanks to olasd for proposing several improvements over the initial
submission.

Addresses swh-alter#4