winery rbd image mapper: handle already mapped shards
Introduce a `image_mapped` function in the Pool, that checks whether an
image is mapped, and if so, if it's mapped read-only or read-write.

This uses file permissions on the image path to determine which mode the
shard is mapped into. For this to work, the udev rules must be applied
like so:

  KERNEL=="rbd[0-9]*", ENV{DEVTYPE}=="disk", ACTION=="add", ATTR{device/pool}=="*shards", ATTR{ro}=="0", GROUP="winery-writer", MODE="0664"
  KERNEL=="rbd[0-9]*", ENV{DEVTYPE}=="disk", ACTION=="add", ATTR{device/pool}=="*shards", ATTR{ro}!="0", GROUP="winery-writer", MODE="0444"
  KERNEL=="rbd[0-9]*", ENV{DEVTYPE}=="disk", ACTION=="add", ATTR{device/pool}=="test-winery-ceph-pool", ATTR{ro}=="0", GROUP="winery-writer", MODE="0664"
  KERNEL=="rbd[0-9]*", ENV{DEVTYPE}=="disk", ACTION=="add", ATTR{device/pool}=="test-winery-ceph-pool", ATTR{ro}!="0", GROUP="winery-writer", MODE="0444"