Implement a file-backed image pool for Winery tests
As an alternative to running the tests with a Ceph backend, we
implement an alternative image pool based on simple files.
We represent unmapped images by setting their permissions to 0o000.

The tests are modified to run with both the file-backed image pool
and the Ceph pool. Those latter ones will be skipped if Ceph is not
available.