winery: remove confusing cursor passing logic in object deletion
The transaction state of the PostgreSQL connection is attached to the
db-api connection object, not to the cursor (which is just a lightweight
object to represent the back-and-forth between a query and the results
of the queries). To share the transaction state, across functions, when
using connection pooling, we need to share the connection object.

In that specific case, the db attribute of the SharedBase object is
already reused between calls and holds the transaction state, so the
cursor passing is unnecessary.