Remove empty __init__.py files
Especially the main one in swh/objstorage needs to be removed since this
later is a namespace, in addition to being a package, in which
swh-objstorage-replayer installs itself. So for the namespace importing
mechanism to work properly, it must be an actual namespace (handled by
the NamespaceLoader, not a package backed by the __init__ file).