winery: release RW shards to STANDBY after an idle timeout delay
We run winery writers in active/active mode on two hosts, but only
usually route requests to the host which is currently managing the
primary database. When the writer fails over, shards would keep being
locked in WRITING mode, even though they're idle.

Implement a RW shard idle timeout with a watchdog thread which gets
pinged every time an object is added. The timeout defaults to 5 minutes,
but can be adjusted with a configuration variable.