Add statsd probing in main methods (add, get and __contains__)
Add a simple timing statsd probe for these methods.

For MultiplexerObjStorage.get(), also send a counter metrics
related to which backend did returned the object.

The counter have add a 'backend' tag with the name of the backend
that returned the object, and a 'backend_number' tag so it's easier
to figure how many fallbacks had to be reached to actually
retrieve a content object.

Note that it does not report timing probes for each backend, since
these metrics are already reported by the backend themselves.