multiplexer: Check backends with respect to their read-only nature
Prior to this commit, a multiplexer instance would start during its
initialization to check the configuration of all its backends but also check
whether they can write or not even when not needed. As it exists read-only
backends for a multiplexer, that would just fail to start the service.

This evolves the objstorage to grow a check_write attribute (hence parametric
through configuration file). And then adapts accordingly the multiplexer
constructor so it can check the configuration of its backends according to
their respective configuration.

Refs. swh/devel/swh-objstorage#4746
Refs. swh/infra/sysadm-environment#5334