Improve support for non-sha1 PRIMARY_HASH
The main goal of this commit is adding support for PRIMARY_HASH==sha256
in the default implementation of list_content. To do so, we need to
introduce PRIMARY_HASH in the ObjStorageInterface, which pulls a bunch
of changes to appease mypy.

Supposedly this might make it easier to handle configurable PRIMARY_HASH?