Clean up winery benchmark tooling
We don't expect to be using Grid'5000 again so this code has not been
maintained with expectations that it would work again. This allows us to
streamline future refactorings as well.

A future objstorage benchmark tool could take inspiration from a lot of
this code, but be agnostic to the backend used.