winery: run deletion tests on both pool backends
To run these tests on the ceph backend, we actually need to release
shards so they can be unmapped (the file-based backend can just do
whatever to a file that is being held open, but a block device needs
more care). The newly run test also unearthed a small file descriptor
leak in the swh.perfecthash object deletion paths.