pypy-compat: transform memory view into bytes in RWShard.all()
postgresql returns `bytes` for `bytea` only for implementation using
`ctypes`. The PyPy interpreteur use something else (likely use `cffi`),
and still return memory-view. So we turn them into bytes to match the
documented code and the expectation of the rest of the code (especially
the compiled part that explicitly checks for `bytes`).